refactor touch data handling to use double for precision and update related classes

+++ lib/page/touchscreen_page.dart
@@ -78,8 +78,8 @@ class TouchScreen extends HookConsumerWidget {
           },
           onPointerMove: (event) {
             udpSenderService.setSingleTouchData(
-              x: event.position.dx.toInt(),
-              y: event.position.dy.toInt(),
+              x: event.position.dx,
+              y: event.position.dy,
             );
           },
         ),
